refactor(solver): keep generic inference helpers under size cap
