+++ crates/tsz-solver/src/operations/generic_call/inference_helpers.rs
@@ -6,20 +6,6 @@ use crate::operations::{AssignabilityChecker, CallEvaluator};
 use crate::types::{FunctionShape, ObjectFlags, ParamInfo, TypeData, TypeId, TypePredicate};
 use rustc_hash::{FxHashMap, FxHashSet};
 
-use super::foreign_param_shapes::is_bare_foreign_type_param;
-
-pub(super) fn is_substantive_inference_candidate(
-    interner: &dyn crate::construction::TypeDatabase,
-    ty: TypeId,
-    local_type_params: &FxHashSet<tsz_common::Atom>,
-    local_placeholders: &[tsz_common::Atom],
-) -> bool {
-    !ty.is_any_unknown_or_error()
-        && !is_bare_foreign_type_param(interner, ty, local_type_params, local_placeholders)
-        && !crate::visitor::contains_type_parameters(interner, ty)
-        && !crate::type_queries::contains_infer_types_db(interner, ty)
-}
-
 impl<'a, C: AssignabilityChecker> CallEvaluator<'a, C> {
     pub(super) fn eval_type_param_default(
         &mut self,
