+++ crates/tsz-solver/src/operations/generic_call/resolve/finalize.rs
@@ -1,7 +1,8 @@
 use crate::inference::infer::{InferenceError, InferenceVar};
 use crate::instantiation::instantiate::{TypeSubstitution, instantiate_type};
-use crate::operations::generic_call::foreign_param_shapes::is_bare_foreign_type_param;
-use crate::operations::generic_call::inference_helpers::is_substantive_inference_candidate;
+use crate::operations::generic_call::foreign_param_shapes::{
+    is_bare_foreign_type_param, is_substantive_inference_candidate,
+};
 use crate::operations::generic_call::readonly_direct_inference;
 use crate::operations::generic_call::{
     constraint_contains_primitive_constrained_type_param,
